feat: add weather module
